feat: ransform the TreatmentPhaseEdit page into a component and place it on the TreatmentPhase page
