+++ src/pages/treatmentPhases/TreatmentPhases.tsx
@@ -10,7 +10,10 @@ import DefaultPatientPage from "../../components/defaultPatientPage/DefaultPatie
 import Button from "../../components/button/Button";
 import ModalTreatmentPhase from "../../components/modalTreatmentPhase/ModalTreatmentPhase";
 import Phase from "./phase/Phase";
+import TreatmentPhaseEdit from "./treatmentPhaseEdit/TreatmentPhaseEdit";
 import { BoxButton, Main, PhaseBlock, PhaseTitle, Section, Title } from "./TreatmentPhases.styles";
+
+const phase: PhaseType = { active: true, frequency: '7 dias', endTreatment: '2024-04-30', dosage: '1:10.000', id: "1", phaseNumber: 1, startTreatment: '2024-03-15' }
 
 const TreatmentPhases = () => {
   const { id: idPatient } = useParams();
@@ -36,6 +39,7 @@ const TreatmentPhases = () => {
   const [modal, setModal] = useState(false);
   const [phaseSelected, setPhaseSelected] = useState<PhaseType | null>(null);
   const [phaseProgress, setPhaseProgress] = useState<number>(0);
+  const [phaseEdit, setPhaseEdit] = useState<PhaseType | null>(null);
 
   const selectPhase = (phase: PhaseType) => {
     const newPhase = phase === phaseSelected ? null : phase;
@@ -49,7 +53,7 @@ const TreatmentPhases = () => {
     if (!phaseSelected) return;
 
     const token = getToken();
-    if (confirm('Tem certeza que deseja finalizar essa fase?')) {
+    if (confirm('Deseja finalizar essa fase?')) {
       const { url, headers } = userRequest.PUT_PHASE_STATUS_BY_ID(phaseSelected.id, token);
 
       const body = {
@@ -102,7 +106,7 @@ const TreatmentPhases = () => {
   const showPhase = (phase: PhaseType) => {
     if (phaseSelected?.phaseNumber === phase.phaseNumber) {
       return (
-        <Phase phase={phase} progress={phaseProgress} />
+        <Phase phase={phase} progress={phaseProgress} setPhase={setPhaseEdit} />
       );
     }
   };
@@ -124,24 +128,26 @@ const TreatmentPhases = () => {
   return (
     <Main>
       <DefaultPatientPage patient={patient} >
-        <Section>
-          <Title>Fases</Title>
-          {getPhases()}
-          <BoxButton className={hasPhases() ? '' : 'centered'}>
-            {
-              hasPhases() &&
-              <Button
-                disabled={!phaseSelected?.active}
-                onClick={finishPhase}
-              >
-                Finalizar Fase
-              </Button>
-            }
-            <Button onClick={() => setModal(!modal)}>Adicionar Fase</Button>
-          </BoxButton>
-        </Section>
-        {modal && <ModalTreatmentPhase setModal={setModal} />}
+        {phaseEdit ?
+          <TreatmentPhaseEdit phase={phaseEdit} setPhase={setPhaseEdit} /> :
+          <Section>
+            <Title>Fases</Title>
+            {getPhases()}
+            <BoxButton className={hasPhases() ? '' : 'centered'}>
+              {
+                hasPhases() &&
+                <Button
+                  disabled={!phaseSelected?.active}
+                  onClick={finishPhase}
+                >
+                  Finalizar Fase
+                </Button>
+              }
+              <Button onClick={() => setModal(!modal)}>Adicionar Fase</Button>
+            </BoxButton>
+          </Section>}
       </DefaultPatientPage>
+      {modal && <ModalTreatmentPhase setModal={setModal} />}
     </Main>
   );
 };
