+++ src/router/Router.tsx
@@ -7,7 +7,6 @@ import MenuMedico from "../pages/menuMedico/MenuMedicoPage";
 import PerfilPaciente from "../pages/perfilPaciente/PerfilPacientePage";
 import TreatmentPhases from "../pages/treatmentPhases/TreatmentPhases";
 import VacinasPaciente from "../pages/vacinasPaciente/VacinasPaciente";
-import TreatmentPhaseEdit from "../pages/treatmentPhaseEdit/TreatmentPhaseEdit";
 
 const Router = () => {
   return (
@@ -19,7 +18,6 @@ const Router = () => {
           <Route path="cadastro" element={<Cadastro />} />
           <Route path="menu-medico" element={<MenuMedico />} />
           <Route path="/paciente/perfil/:id" element={<PerfilPaciente />} />
-          <Route path="/paciente/fase/edit" element={<TreatmentPhaseEdit />} />
           <Route path="/paciente/tratamento/:id" element={<TreatmentPhases />} />
           <Route path="/paciente/vacinas/:id" element={<VacinasPaciente />} />
         </Routes>
